kirkwood u-boot: initialize PHY and set board ID
When changing machine type to hikirk, several commands
that were only executed conditionally (dependent on
openrd ultimate) were not executed anymore.

This resulted in a failure during network initialization.

This patch makes sure all of these commands are executed
again. This repairs network and may have other effects
that we are not aware of yet.

Signed-off-by: Angelika Tobisch <tobisch@dresearch-fe.de>
Signed-off-by: Mario Schuknecht <schuknecht@dresearch-fe.de>