u-boot: separate config file
Tool mkimage generates from a config file a u-boot header. A config file is
required for each u-boot header (spi, sata, uart, nand). The config file
consits of a header area and a data area. The data area is the same for all
u-boot headers.
In order not to keep the data area multiple times, the config files are
separated in a header file and a data file. And the config files are created
during the build process.

Signed-off-by: Mario Schuknecht <mario.schuknecht@dresearch-fe.de>