hikirk.conf: set virtual/libgl preferred provider
Due to dependancies the packages 'mesa' and 'mesa-gl' will be built during
'bitbake world'. Both packages provide target 'virtual/libgl'.
During 'bitbake world' an error occurs with these packages.
This error prevents the built of targets for MACHINE hikirk. The cause is not
known, there are no dependencies. As workaround, the built of package
'mesa-gl' is prevented by changing the PREFERED_PROVIDER for 'virtual/libgl'.
Default is 'mesa-gl'.

Signed-off-by: Mario Schuknecht <mario.schuknecht@dresearch-fe.de>